requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.1.0
langgraph>=0.2.0
langchain-openai>=0.2.0
faiss-cpu>=1.8
pydantic>=2.0
python-dotenv>=1.0
